ufbx_math/libc cleanup
